Fix #5298 and an unrelated bug (#5322)
* Fix #5298

* Fix an issue with viewing JobResults missing a date_done value

+++ nautobot/extras/signals.py
@@ -20,6 +20,7 @@ from django.utils import timezone
 from django_prometheus.models import model_deletes, model_inserts, model_updates
 
 from nautobot.core.celery import app, import_jobs_as_celery_tasks
+from nautobot.core.models import BaseModel
 from nautobot.core.utils.config import get_settings_or_config
 from nautobot.core.utils.logging import sanitize
 from nautobot.extras.choices import JobResultStatusChoices, ObjectChangeActionChoices
@@ -167,10 +168,12 @@ def _handle_deleted_object(sender, instance, **kwargs):
     if change_context_state.get() is None:
         return
 
-    associations = ContactAssociation.objects.filter(
-        associated_object_type=ContentType.objects.get_for_model(type(instance)), associated_object_id=instance.pk
-    )
-    associations.delete()
+    if isinstance(instance, BaseModel):
+        associations = ContactAssociation.objects.filter(
+            associated_object_type=ContentType.objects.get_for_model(type(instance)), associated_object_id=instance.pk
+        )
+        associations.delete()
+
     if hasattr(instance, "notes") and isinstance(instance.notes, NotesQuerySet):
         notes = instance.notes
         notes.delete()
